Creates testing subpackage
This package houses a pytest_generate_tests function that
enables someone to easily write tests like this:

```python
import pytest
from burr.testing import pytest_generate_tests  # noqa: F401
# TODO: import action you're testing, i.e. import {ACTION_NAME}.

@pytest.mark.file_name("{FILE_NAME}")
def test_{ACTION_NAME}(input_state, expected_state):
    \"\"\"Function for testing the action\"\"\"
    input_state = state.State(input_state)
    expected_state = state.State(expected_state)
    _, output_state = {ACTION_NAME}(input_state)  # exercise the action
    assert output_state == expected_state
```
i.e. one can create files that map to tests cases for an action.
The expected state also is stored in the file.
